### 2025-04-23 #7880 @elbeejay
@davemlz would you be able to review this submission to JOSS titled "RadGEEToolbox"? I recall you publishing a GEE+Python tool in JOSS in the past (`eemont` I believe). The JOSS review process has not changed radically since then, please let me know if you're able and interested in being on the other side of peer-reviewing this time!
