### 2025-04-23 #7880 @elbeejay
@2320sharon I saw your recent JOSS publication on `CoastSeg` and wanted to ask if you'd be open to reviewing `RadGEEToolbox` for JOSS. Please let me know if you are able and interested!
