### 2025-04-24 #7880 @JohnVolk
Yes @elbeejay, I would be happy to review this paper and software. It looks quite interesting!
