+++ paper/paper.bib
@@ -42,7 +42,8 @@
 @article{Radwin&Bowen:2024,
     url = {https://giw.utahgeology.org/giw/index.php/geosites/article/view/134},
     Author = {M. {Radwin} and B. {Bowen}},
-    Journal = {Utah Geological Association},
+    Doi = {https://doi.org/10.31711/ugap.v51i.134},
+    Journal = {Geosites},
     Title = {Evolution of Great Salt Lake’s Exposed Lakebed (1984-2023): Variations in Sediment Composition, Water, and Vegetation from Landsat OLI and Sentinel MSI Satellite Reflectance Data},
     Volume = {51},
     Year = {2024}
