+++ paper/paper.md
@@ -14,97 +14,52 @@ authors:
 affiliations:
  - name: Dept. of Geology & Geophysics, University of Utah, United States
    index: 1
-date: 3 March 2025
+date: 24 April 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-`RadGEEToolbox` is an open-source Python package that facilitates the processing and analysis of both multispectral and Synthetic Aperture Radar (SAR) satellite imagery using `Google Earth Engine` [@Gorelick:2017], including Landsat TM & OLI (courtesy of U.S. Geological Survey), Sentinel-1 SAR (courtesy of European Space Agency), and Sentinel-2 MSI (courtesy of European Space Agency) datasets. By providing an intuitive, modular framework, it simplifies complex remote sensing workflows, enabling researchers to efficiently manage, process, and analyze satellite imagery across a range of environmental and geospatial applications. The package is designed for time series analyses of large-scale remote sensing datasets, reducing the amount of custom scripting required in `Google Earth Engine` while ensuring reproducibility and efficiency. Researchers working in landscape change, environmental monitoring, natural hazards, and land cover classification can benefit from `RadGEEToolbox`’s ability to streamline workflows from data acquisition to final analysis.
+`RadGEEToolbox` is an open-source Python package for processing and analyzing multispectral and Synthetic Aperture Radar (SAR) imagery using `Google Earth Engine` [@Gorelick:2017]. It supports Landsat TM/OLI, Sentinel-1 SAR, and Sentinel-2 MSI datasets, providing a modular, streamlined workflow for landscape change detection, environmental monitoring, and land classification. It reduces the need for repetitive code while promoting reproducibility and efficiency.
 
-The package includes specialized tools for both multispectral and SAR data processing, offering capabilities such as image subsetting, mosaic generation, cloud/water masking, spectral index calculations, speckle filtering, multilooking, backscatter transformations, transect extractions, spatial statistics, and more. One of the advantages of `RadGEEToolbox` is its ability to seamlessly integrate user-defined `Google Earth Engine` functions or operations, allowing researchers to modify and extend their analyses without being restricted to predefined operations offered by this package. `RadGEEToolbox` also includes preconfigured visualization parameters, facilitating intuitive exploration of geospatial datasets.
+Features include subsetting, mosaicking, cloud/water masking, spectral index calculations, speckle filtering, backscatter conversions, spatial statistics, and integration of custom Earth Engine functions.
 
+# Statement of Need
 
-# Statement of need
-
-Processing and analyzing remote sensing data in `Google Earth Engine` requires significant programming expertise, often leading to steep learning curves for new users and complex, bloated code. While `Google Earth Engine` provides a powerful API for accessing and manipulating Earth observation datasets, performing even routine tasks such as filtering image collections, computing spectral indices, or transforming SAR backscatter values often requires extensive scripting and re-use of the same functions or code across multiple projects. `RadGEEToolbox` addresses these issues by streamlining and automating many of these repetitive and complex tasks while still allowing advanced users the flexibility to customize their analyses. `RadGEEToolbox` is designed for students, researchers, remote sensing analysts, and environmental scientists at any level of experience who are working with Landsat, Sentinel-1, or Sentinel-2 datasets. Ultimately, `RadGEEToolbox` makes it easier to perform multi-temporal analyses, extract surface properties or time series, generate land cover statistics, and visualize key environmental indicators by handling much of the complexity inherent in `Google Earth Engine` API interactions and image processing tasks.
[…]
