### 2025-04-24 #7880 @radwinskis
I have updated the `paper.md` file to be less than 1000 words and believe formatting to be correct.

Thank you @elbeejay for handling the editing. For another reviewer, https://github.com/giswqs could be excellent if they are willing.
