### 2025-04-26 #7880 @davemlz
> [@davemlz](https://github.com/davemlz) would you be able to review this submission to JOSS titled "RadGEEToolbox"? I recall you publishing a GEE+Python tool in JOSS in the past (`eemont` I believe). The JOSS review process has not changed radically since then, please let me know if you're able and interested in being on the other side of peer-reviewing this time!

Hi @elbeejay! Thank you for considering me for reviewing this submission. However, I will be unavailable to do so in the next weeks. But I would be happy to do reviews in the future!
