### 2025-04-27 #7880 @elbeejay
@McWhity would you be interested in reviewing this submission, `RadGEEToolbox`, for JOSS? I realize you are an expert in SAR analysis and have prior experience both publishing and reviewing for JOSS, so I thought you'd be good candidate to review this submission. Please consider it and let us know, thanks!
