### 2025-04-27 #7880 @elbeejay
@lwasser will you consider reviewing this submission, `RadGEEToolbox`, for JOSS? Your prior publication on `EarthPy` and expertise in Python for earth science make you a strong potential reviewer. Please let me know if you are interested and able in helping us out with this review.
