### 2025-04-29 #7880 @McWhity
@elbeejay it looks like a nice submission. I would be willing to do the review. However, I currently have my own submission under review at JOSS, and I’m behind on completing a comprehensive update. I’d prefer to complete that update before starting this review. I expect to finish it by next Thursday, May 8th. So, I’m available to do the review, but I can only begin after May 8th.
