### 2025-04-30 #7880 @elbeejay
That would work just fine, thanks @McWhity - appreciate the transparency!
