### 2025-05-05 #8164 @elbeejay
Hi @JohnVolk and @McWhity. Thanks again for agreeing to review `RadGEEToolbox`. In the top comment of this issue you'll find instructions for creating your own reviewer checklist as well a link to the JOSS reviewer guidelines. Please don't hesitate to reach out to me if you have any questions.

We're asking reviewers to complete their first pass and set of comments within 6 weeks of the review issue beginning, so for us this would be before June 16, 2025. Please just be open and communicate if you're going to be unable to meet that deadline.

Thanks again,
Jay
