### 2025-05-06 #7880 @sebhaan
> [@sebhaan](https://github.com/sebhaan) would you consider reviewing `RadGEEToolbox` for JOSS? Your experience developing and publishing `Geodata-Harvester` make you a great potential peer-reviewer. Please let me know if you are interested and able to help us out with this review.

Thank you for considering me to review the RadGEEToolbox. It looks like a useful tool for working with remote sensing data, and I'm willing to review the work. I haven't reviewed for JOSS yet, please let me know the next steps in the review process, including the expected timeline.
