### 2025-05-07 #8164 @elbeejay
@sebhaan - details regarding the review process are in the first post of this issue thread as well as in the [JOSS online documentation](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#guiding-principles). But the short of it is that JOSS attempts to have a transparent and open review process, where reviewers and authors can work through comments without being stuck to a highly formal review -> respond -> re-review structure. Instead we have a reviewer checklist that you'll generate and use to guide your review, and you can tick off boxes when review criteria are satisfied.

We are flexible in the review deadline, right now we are asking you to complete your first pass through within 6 weeks. If you need additional time, my only ask is that you just communicate with us so everyone is aware that you'll be working on it and when they might expect to see comments.
