### 2025-05-07 radwinskis/RadGEEToolbox#1 @JohnVolk
Improve online documentation
Hi @radwinskis, as I am starting on the JOSS review for your software and article, one of the first things I noticed was that the online ReadTheDocs documentation seems to be quite limited. There is a bit more on the GitHub README but it is a bit hard to read and follow, and limited in scope. It would benefit the software a lot to improve both, the README and ReadTheDocs documentation. In line with the JOSS requirements, I recommend adding more than the basic API documentation (docstrings), at a minimum adding installation instructions and a tutorial and/or example usage. You may also consider adding other useful sections such as contributing, change log info, automated tests, etc.

In addition, to improve the docstrings and API documentation I suggest adopting Google, Numpy, or reST formating so that Sphinx can properly highlight the documentation on ReadTheDocs. The current format is human-readable but not formatted for Sphinx. To use Google or Numpy you can just add this to your conf.py:

```python
extensions = ['sphinx.ext.napoleon']
```
