### 2025-05-09 radwinskis/RadGEEToolbox#1 @radwinskis
@JohnVolk Thank you for bringing this to my attention. I will work on improving the ReadTheDocs by correcting my docstrings, using Google format, and adding detailed installation/examples info. If you could suggest specific ways I can improve the README that would be greatly appreciated as I thought it had enough information - but I totally understand the need to expand the ReadTheDocs documentation. Thanks again for the helpful recommendations.
