### 2025-05-10 radwinskis/RadGEEToolbox#1 @JohnVolk
Regarding the README, I just recommend to clean it up. There seems to be a lot of unnecessary text and sections that could be streamlined. Also, the example use formatting has a lot of text and could benefit from a visual or two. It's good to keep it shorter and simple. Lots of bold font and emoji's are a bit distracting to me but that's a personal preference and fine in some cases.
