### 2025-05-14 97cdb5f
reformatted docstrings and ReadTheDocs documentation. Updates to examples and readme.
