+++ RadGEEToolbox/LandsatCollection.py
@@ -5,7 +5,9 @@ class LandsatCollection:
     """
     Class object representing a combined collection of NASA/USGS Landsat 5, 8, and 9 TM & OLI surface reflectance satellite images at 30 m/px from Google Earth Engine.
 
-    This class provides methods to filter, process, and analyze Landsat satellite imagery for a given period and region
+    This class provides methods to filter, process, and analyze Landsat satellite imagery for a given period and region.
+
+    Inspect the documentation or source code for details on the methods and properties available.
 
     Arguments:
         start_date (str): Start date string in format of yyyy-mm-dd for filtering collection (required unless collection is provided)
@@ -289,10 +291,10 @@ class LandsatCollection:
         Adds date to image properties as 'Date_Filter'.
 
         Args: 
-        image (ee.Image): Input image
+            image (ee.Image): Input image
 
         Returns: 
-        image (ee.Image): Image with date in properties.
+            ee.Image: Image with date in properties.
         """
         date = ee.Number(image.date().format('YYYY-MM-dd'))
         return image.set({'Date_Filter': date})
@@ -300,30 +302,30 @@ class LandsatCollection:
     @staticmethod
     def landsat5bandrename(img):
         """
-        Function to rename Landsat 5 bands to match Landsat 8 & 9.
+        Renames Landsat 5 bands to match Landsat 8 & 9.
 
         Args: 
-        image (ee.Image): input image
+            image (ee.Image): input image
         
         Returns: 
-        image (ee.Image): image with renamed bands
+            ee.Image: image with renamed bands
         """
         return img.select('SR_B1', 'SR_B2', 'SR_B3', 'SR_B4', 'SR_B5', 'SR_B7', 'QA_PIXEL').rename('SR_B2', 'SR_B3', 'SR_B4', 'SR_B5', 'SR_B6', 'SR_B7', 'QA_PIXEL')
     
     @staticmethod
     def landsat_ndwi_fn(image, threshold, ng_threshold=None):
         """
-        Function to calculate ndwi from GREEN and NIR bands (McFeeters, 1996 - https://doi.org/10.1080/01431169608948714) for Landsat imagery and mask image based on threshold. 
-        Can specify separate thresholds for Landsat 5 vs 8 & 9 images, where the threshold 
-        argument applies to Landsat 5 and the ng_threshold argument applies to Landsat 8 & 9.
+        Calculates ndwi from GREEN and NIR bands (McFeeters, 1996 - https://doi.org/10.1080/01431169608948714) for Landsat imagery and mask image based on threshold. 
+        
+        Can specify separate thresholds for Landsat 5 vs 8 & 9 images, where the threshold argument applies to Landsat 5 and the ng_threshold argument applies to Landsat 8 & 9.
 
         Args: 
-        image (ee.Image): input image
-        threshold (int): value between -1 and 1 where pixels less than threshold will be masked, applies to landsat 5 when ng_threshold is also set.
-        ng_threshold (int | optional): integer threshold to be applied to landsat 8 or 9 where pixels less than threshold are masked
+            image (ee.Image): input image
+            threshold (float): value between -1 and 1 where pixels less than threshold will be masked, applies to landsat 5 when ng_threshold is also set.
+            ng_threshold (float, optional): integer threshold to be applied to landsat 8 or 9 where pixels less than threshold are masked
 
         Returns:
-        image (ee.Image): ndwi image
+            ee.Image: ndwi image
         """
         ndwi_calc = image.normalizedDifference(['SR_B3', 'SR_B5']) #green-NIR / green+NIR -- full NDWI image
         water = ndwi_calc.updateMask(ndwi_calc.gte(threshold)).rename('ndwi').copyProperties(image) 
[…]
