+++ RadGEEToolbox/Sentinel1Collection.py
@@ -4,9 +4,11 @@ import pandas as pd
 import numpy as np
 class Sentinel1Collection:
     """
-    Class object representing a collection of Sentinel-1 C-band Synthetic Aperture Radar (SAR) GRD data at 10 m/px resolution from Google Earth Engine (GEE).
+    Class object representing a collection of Sentinel-1 C-band Synthetic Aperture Radar (SAR) GRD data at 10 m/px resolution from Google Earth Engine (GEE). Units of backscatter are in decibels (dB) by default.
 
-    This class module provides methods to filter, process, and analyze Sentinel-1 satellite imagery for a given period and region
+    This class module provides methods to filter, process, and analyze Sentinel-1 satellite imagery for a given period and region.
+
+    Inspect the documentation or source code for details on the methods and properties available.
 
     Arguments:
         start_date (str): Start date string in format of yyyy-mm-dd for filtering collection (required unless collection is provided)
@@ -23,7 +25,7 @@ class Sentinel1Collection:
 
         bands (str or list): Band(s) of interest in each image (optional, must match polarization type; default is ['VV', 'VH'])
 
-        orbit_direction (str or list): Orbit direction for filtering collection. Options: 'ASCENDING' and/or 'DESCENDING' (required unless collection is provided)
+        orbit_direction (str or list): Orbit direction for filtering collection. Options: 'ASCENDING' and/or 'DESCENDING' (required unless collection is provided). For example, ['ASCENDING', 'DESCENDING'] will include both ascending and descending images.
 
         boundary (ee.Geometry): Boundary for filtering images to images that intersect with the boundary shape (optional)
 
@@ -181,7 +183,7 @@ class Sentinel1Collection:
             elif orbit_direction == ['ASCENDING', 'DESCENDING']:
                 self.orbit_direction = orbit_direction
             elif orbit_direction not in ['ASCENDING', 'DESCENDING']:
-                raise ValueError("Orbit direction must be either 'ASCENDING' or 'DESCENDING'")
+                raise ValueError("Orbit direction must be either 'ASCENDING' or 'DESCENDING', or '['ASCENDING', 'DESCENDING']' ")
             else:
                 pass
 
@@ -276,10 +278,10 @@ class Sentinel1Collection:
         Adds date to image properties as 'Date_Filter'.
 
         Args: 
-        image (ee.Image): Input image
+            image (ee.Image): Input image
 
         Returns: 
-        image (ee.Image): Image with date in properties.
+            ee.Image: Image with date in properties.
         """
         date = ee.Number(image.date().format('YYYY-MM-dd'))
         return image.set({'Date_Filter': date})
@@ -292,15 +294,15 @@ class Sentinel1Collection:
         and store the value as image property (matching name of chosen band).
 
         Args:
-        image (ee.Image): input ee.Image
-        band_name (string): name of band (string) for calculating area
-        geometry (ee.Geometry): ee.Geometry object denoting area to clip to for area calculation
-        threshold (int): integer threshold to specify masking of pixels below threshold (defaults to -1)
-        scale (int): integer scale of image resolution (meters) (defaults to 30)
-        maxPixels (int): integer denoting maximum number of pixels for calculations
-        
-        Returns:
-        image (ee.Image): ee.Image with area calculation stored as property matching name of band
+            image (ee.Image): input ee.Image
+            band_name (string): name of band (string) for calculating area
+            geometry (ee.Geometry): ee.Geometry object denoting area to clip to for area calculation
+            threshold (float): integer threshold to specify masking of pixels below threshold (defaults to -1)
+            scale (int): integer scale of image resolution (meters) (defaults to 30)
+            maxPixels (int): integer denoting maximum number of pixels for calculations
+            
+        Returns: 
[…]
