### 2025-05-14 radwinskis/RadGEEToolbox#1 @radwinskis
@JohnVolk I have implemented significant changes to the README.md, updated all the docstrings to correct formatting, added multiple sections to the RTD docs, and have updated the example notebooks to provide more examples of how to use RadGEEToolbox. Let me know if there are any other immediate issues I should address, and if not I will close this issue. Thank you for helping me make these significant improvements!
