### 2025-05-14 ecef80e
fixing wording issue in paper and function docstrings in Sentinel1Collection
