### 2025-05-14 82a3e07
Image descriptions added to readme
