### 2025-05-16 radwinskis/RadGEEToolbox#1 @JohnVolk
Hi @radwinskis, this is a big improvement particularly the example usage page with some visual results on ReadTheDocs- that helps a lot. The class methods docstrings are being parsed correctly by Sphinx now. The README is a lot cleaner as well. However, there are still some issues I am seeing overall with the documentation that should help improve the software:

The class docstrings are not rendered properly in the ReadTheDocs page—parameters and sections like Args: and Returns: appear as plain text, indicating that Sphinx formatting (e.g., via the napoleon extension) is not being applied. The docstring includes a long unparsed list of attributes and methods that should be handled by autodoc instead. Usage examples and formatting should follow Google or NumPy style to be correctly interpreted. You may need to wrap the code in an rst directive for example.

For better readability and structure of ReadTheDocs I would put all the class and methods API docs under one section on the page instead of many listed out on the website, for example, "LandsatCollection Documentation", etc. should all be under API reference as subsections, and I would drop the word "Documentation" that follows each submodule/class.

Also, the example notebooks show large print outs of the returned data from difference EE methods. I would limit those to only one per code block, and add a bit more of an explanation of what is being printed.
