### 2025-05-22 e406a3f
class docstrings fixes, enhanced example notebooks, and expanded documentation
