+++ RadGEEToolbox/LandsatCollection.py
@@ -3,221 +3,50 @@ import pandas as pd
 import numpy as np
 class LandsatCollection:
     """
-    Class object representing a combined collection of NASA/USGS Landsat 5, 8, and 9 TM & OLI surface reflectance satellite images at 30 m/px from Google Earth Engine.
+    Represents a user-defined collection of NASA/USGS Landsat 5, 8, and 9 TM & OLI surface reflectance satellite images at 30 m/px from Google Earth Engine (GEE).
 
-    This class provides methods to filter, process, and analyze Landsat satellite imagery for a given period and region.
+    This class enables simplified definition, filtering, masking, and processing of multispectral Landsat imagery.
+    It supports multiple spatial and temporal filters, caching for efficient computation, and direct computation of 
+    key spectral indices like NDWI, NDVI, halite index, and more. It also includes utilities for cloud masking,
+    mosaicking, zonal statistics, and transect analysis.
+
+    Initialization can be done by providing filtering parameters or directly passing in a pre-filtered GEE collection.
 
     Inspect the documentation or source code for details on the methods and properties available.
 
-    Arguments:
-        start_date (str): Start date string in format of yyyy-mm-dd for filtering collection (required unless collection is provided)
-
-        end_date (str): End date string in format of yyyy-mm-dd for filtering collection (required unless collection is provided)
-
-        tile_row (int or list): WRS-2 tile row of Landsat image (required unless boundary or collection is provided) - see https://www.usgs.gov/landsat-missions/landsat-shapefiles-and-kml-files
-
-        tile_path (int or list): WRS-2 tile path of Landsat image (required unless boundary or collection is provided) - see https://www.usgs.gov/landsat-missions/landsat-shapefiles-and-kml-files
-
-        cloud_percentage_threshold (int): Integer percentage threshold where only imagery with cloud % less than threshold will be provided (defaults to 100)
-
-        boundary (ee.Geometry): Boundary for filtering images to images that intersect with the boundary shape (optional) - will override tile specifications
-        
-        collection (ee.ImageCollection): Optional argument to provide an ee.ImageCollection object to be converted to a LandsatCollection object - will override other arguments!
+    Args:
+        start_date (str): Start date in 'YYYY-MM-DD' format. Required unless `collection` is provided.
+        end_date (str): End date in 'YYYY-MM-DD' format. Required unless `collection` is provided.
+        tile_row (int or list of int): WRS-2 tile row(s) to filter by. Ignored if `boundary` or `collection` is provided. See https://www.usgs.gov/landsat-missions/landsat-shapefiles-and-kml-files
+        tile_path (int or list of int): WRS-2 tile row(s) to filter by. Ignored if `boundary` or `collection` is provided. See https://www.usgs.gov/landsat-missions/landsat-shapefiles-and-kml-files
+        cloud_percentage_threshold (int, optional): Max allowed cloud cover percentage. Defaults to 100.
+        boundary (ee.Geometry, optional): A geometry for filtering to images that intersect with the boundary shape. Overrides `tile_path` and `tile_row` if provided.
+        collection (ee.ImageCollection, optional): A pre-filtered Landsat ee.ImageCollection object to be converted to a LandsatCollection object. Overrides all other filters.
 
     Attributes:
-        collection: Returns an ee.ImageCollection object from any LandsatCollection image collection object
+        collection (ee.ImageCollection): The filtered or user-supplied image collection converted to an ee.ImageCollection object.        
         
-        _dates_list: Cache storage for dates_list property attribute
+    Raises:
[…]
