+++ RadGEEToolbox/Sentinel2Collection.py
@@ -3,214 +3,52 @@ import pandas as pd
 import numpy as np
 class Sentinel2Collection:
     """
-    Class object representing a collection of ESA Sentinel-2 MSI surface reflectance satellite images at 10 m/px resolution from Google Earth Engine.
+    Represents a user-defined collection of ESA Sentinel-2 MSI surface reflectance satellite images at 10 m/px from Google Earth Engine (GEE).
 
-    This class provides methods to filter, process, and analyze Sentinel-2 satellite imagery for a given period and region. 
+    This class enables simplified definition, filtering, masking, and processing of multispectral Sentinel-2 imagery.
+    It supports multiple spatial and temporal filters, caching for efficient computation, and direct computation of 
+    key spectral indices like NDWI, NDVI, halite index, and more. It also includes utilities for cloud masking,
+    mosaicking, zonal statistics, and transect analysis.
+
+    Initialization can be done by providing filtering parameters or directly passing in a pre-filtered GEE collection.
 
     Inspect the documentation or source code for details on the methods and properties available.
 
-    Arguments:
-        start_date (str): Start date string in format of yyyy-mm-dd for filtering collection (required unless collection is provided)
-        
-        end_date (str): End date string in format of yyyy-mm-dd for filtering collection (required unless collection is provided)
-        
-        tile (str or list): MGRS tile(s) of Sentinel image (required unless boundary, relative_orbit_number, or collection is provided) | user is allowed to provide multiple tiles as list (note tile specifications will override boundary or orbits) - see https://hls.gsfc.nasa.gov/products-description/tiling-system/
-        
-        cloud_percentage_threshold (int): Integer percentage threshold where only imagery with cloud % less than threshold will be provided (defaults to 100)
-
-        nodata_threshold (int): Integer percentage threshold where only imagery with nodata pixels encompassing a % less than the threshold will be provided (defaults to 100)
-
-        boundary (ee.Geometry): Boundary for filtering images to images that intersect with the boundary shape (optional) - can be used in conjunction with relative_orbit_number
-        
-        relative_orbit_number (int or list): Relative orbit number(s) to filter collection (optional) - can be used in conjunction with boundary | provide multiple values as list
-        
-        collection (ee.ImageCollection): Optional argument to convert an ee.ImageCollection object to a Sentinel2Collection object - will override other arguments!
+    Args:
+        start_date (str): Start date in 'YYYY-MM-DD' format. Required unless `collection` is provided.
+        end_date (str): End date in 'YYYY-MM-DD' format. Required unless `collection` is provided.
+        tile (str or list): MGRS tile(s) of Sentinel image. Required unless `boundary`, `relative_orbit_number`, or `collection` is provided. The user is allowed to provide multiple tiles as list (note tile specifications will override boundary or orbits). See https://hls.gsfc.nasa.gov/products-description/tiling-system/       
+        cloud_percentage_threshold (int, optional): Max allowed cloud cover percentage. Defaults to 100.
+        nodata_threshold (int, optional): Integer percentage threshold where only imagery with nodata pixels encompassing a % less than the threshold will be provided (defaults to 100)
+        boundary (ee.Geometry, optional): A geometry for filtering to images that intersect with the boundary shape. Overrides `tile` if provided.       
+        relative_orbit_number (int or list, optional): Relative orbit number(s) to filter collection. Provide multiple values as list 
[…]
