+++ docs/example_notebooks.rst
@@ -1,0 +1,13 @@
+Examples
+========
+
+To assist in getting you off the ground with using RadGEEToolbox, we have provided a number of example notebooks and will be adding more over time. 
+These notebooks are designed to demonstrate the functionality of RadGEEToolbox and provide a starting point for your own analyses.
+
+.. toctree::
+   :maxdepth: 1
+
+   examples/Image_Collections_&_Useful_Operations
+   examples/Multispectral_Functions_Examples
+   examples/Palettes_and_Visualization
+   examples/S1_SAR_Backscatter_Basic_Usage
