+++ docs/Sentinel2Collection.rst
@@ -1,5 +1,7 @@
-Sentinel2Collection Documentation
----------------------------------
+Sentinel2Collection
+===================
 
 .. automodule:: RadGEEToolbox.Sentinel2Collection
    :members:
+   :undoc-members:
+   :show-inheritance:
