+++ docs/api_reference.rst
@@ -1,0 +1,14 @@
+API Reference
+=============
+
+This section contains the documentation for the classes available in RadGEEToolbox.
+
+.. toctree::
+   :maxdepth: 2
+
+   LandsatCollection
+   Sentinel1Collection
+   Sentinel2Collection
+   CollectionStitch
+   GetPalette
+   VisParams
