+++ docs/Sentinel1Collection.rst
@@ -1,5 +1,7 @@
-Sentinel1Collection Documentation
----------------------------------
+Sentinel1Collection
+===================
 
 .. automodule:: RadGEEToolbox.Sentinel1Collection
    :members:
+   :undoc-members:
+   :show-inheritance:
