+++ docs/LandsatCollection.rst
@@ -1,5 +1,7 @@
-LandsatCollection Documentation
--------------------------------
+LandsatCollection 
+=================
 
 .. automodule:: RadGEEToolbox.LandsatCollection
    :members:
+   :undoc-members:
+   :show-inheritance:

[…]
