### 2025-05-22 radwinskis/RadGEEToolbox#1 @radwinskis
Hi @JohnVolk - thank you for your additional feedback, it is greatly appreciated and insightful! I have now updated all of the class docstrings to render properly and have removed the long list of attributes and methods. Tried to make sure it is now following standard formatting and elements within the class docstring.

For the ReadTheDocs, I have included a section for API Reference which includes all of the API docs and better organizes the documentation.

Regarding the example notebooks, I have updated those to be better organized and more explicit with the walkthrough explanations. I also split print blocks to new cells. There is now a section including the rendered notebooks on the ReadTheDocs documentation as well.

I appreciate your patience and helpful feedback!
