### 2025-05-27 radwinskis/RadGEEToolbox#1 @JohnVolk
Hi @radwinskis, this looks like a huge improvement for the API docstrings, it seems like they are all rendering correctly. My only other suggestion maybe would be to add a bit more visual or numerical (e.g. a table or dataframe print out or just values) output from some of the examples so that the user can see what the results or data may look like. As of now there is the one map and time series plot used in the README and the example, the other notebooks do not have much to view. Otherwise it looks pretty good to me!
