### 2025-06-03 radwinskis/RadGEEToolbox#2 @McWhity
Installation
Hi @radwinskis , I found some issues with your installation guide.

1. Different version numbers in [README](https://github.com/radwinskis/RadGEEToolbox/blob/main/README.md) and [ReadTheDocs](https://radgeetoolbox.readthedocs.io/en/latest/installation.html)
2. Installation via pip does work, but `python -c "import RadGEEToolbox; print(RadGEEToolbox.__version__)"` result in `AttributeError: module 'RadGEEToolbox' has no attribute 'version'`
3. Installation via conda

   - `conda install conda-forge::radgeetoolbox` result in installing radgeetoolbox version 1.5.0 in which module ee is not installed
   - if using python version 3.12 with `conda install conda-forge::radgeetoolbox==1.6.2` the installation fails

```Channels:
 - conda-forge
Platform: linux-64
Solving environment: \ warning  libmamba Added empty dependency for problem type SOLVER_RULE_UPDATE
failed

LibMambaUnsatisfiableError: Encountered problems while solving:
  - package radgeetoolbox-1.6.2-pyhd8ed1ab_0 requires python >=3.6,<3.12, but none of the providers can be installed

Could not solve for environment specs
The following packages are incompatible
├─ pin-1 is installable and it requires
│  └─ python 3.12.* , which can be installed;
└─ radgeetoolbox 1.6.2  is not installable because it requires
   └─ python >=3.6,<3.12 , which conflicts with any installable versions previously reported.
```

4. Manual installation from source does work, but the version problem described above (pip installation) still exists
5. Use/Initialization of GoogleEarthEngine

   -  You mention that the package is designed for both new and advanced users. New users might have some problems to get the package running as they might not be able to connect/initialize to the GoogleEarthEngine
   - `ee.Initialize()` does usually not work for me. I need to use `ee.Initialize(project="your-project-id`)`.
   - Furthermore, I created a new GoogleEarthEngine project for the testing purposes. Thus, I had to enable the API option and I had to register the new project for non-commercial usage.
   - As your code highly depends on the GoogleEarthEngine, I think it might be good to provide some links or small description on how to get started with the GoogleEarthEngine (Authentication and Initialization process). Otherwise new users might not be able to use your python package capabilities.
6. Using Jupyter notebooks (e.g. [Palettes_and_Visualization,ipynb](https://github.com/radwinskis/RadGEEToolbox/blob/main/Example%20Notebooks/Palettes_and_Visualization.ipynb)
   - I had to first install `pip install ipykernel` (jupyter notebook usage) and `pip install geemap` as it is not part of the python package itself
   - Package dependencies issues are a common problem. I would suggest and highly appreciate if you could set up a docker container. One advantage for you would be that you can better handle the installation process (define python version, install all required dependencies). An advantage for users would be that after installing of the docker your package and all the juypter notebooks should work out-of-the-box.
