### 2025-06-03 radwinskis/RadGEEToolbox#1 @McWhity
Hi @radwinskis I would like to add some things concerning the [README](https://github.com/radwinskis/RadGEEToolbox/blob/main/README.md). The usage example is nice but,

-  as described in [issue "Installation"](https://github.com/radwinskis/RadGEEToolbox/issues/2), `ee.Initialize()` does not work for me out-of-the-box.
- In "2. Apply a Cloud Mask and Compute NDWI" I was missing a code snipped how to create the sample image below.
- The code snippet within  "3. Calculate Water Area Time Series" has a typo. Line 3: `geometry=study_area),`. **Bracket ")"** after study_area needs to be removed otherwise the code is not working
```
calculate_water_area = cloud_masked_NDWI_collection.PixelAreaSumCollection(
      band_name='ndwi', #specify band to use from collection
      geometry=study_area), #ee.Geometry() of your study area
      threshold=0, #binary classification threshold for unclassified rasters,
      scale=90 #pixel size for zonal statistics
      )
```
- A code snipped to produce [Figure Lake Powell time series](https://github.com/radwinskis/RadGEEToolbox/blob/main/image-4.png) would be also nice to have
