### 2025-06-04 aa4721d
adding to readme, updating images, adding images to example notebooks, and small fixes
