+++ RadGEEToolbox/__init__.py
@@ -1,6 +1,18 @@
+__version__ = "1.6.3"
+
 from .CollectionStitch import CollectionStitch, MosaicByDate
 from .GetPalette import get_palette
 from .LandsatCollection import LandsatCollection
 from .Sentinel1Collection import Sentinel1Collection
 from .Sentinel2Collection import Sentinel2Collection
 from .VisParams import get_visualization_params
+
+__all__ = [
+    "CollectionStitch",
+    "MosaicByDate",
+    "get_palette",
+    "LandsatCollection",
+    "Sentinel1Collection",
+    "Sentinel2Collection",
+    "get_visualization_params"
+]
