+++ docs/installation.rst
@@ -11,14 +11,14 @@ Prerequisites
 Installing via pip
 ------------------
 
-To install ``RadGEEToolbox`` version 1.6.1 using pip (recommended: use a virtual environment)::
+To install ``RadGEEToolbox`` version 1.6.3 using pip (recommended: use a virtual environment)::
 
-    pip install RadGEEToolbox==1.6.1
+    pip install RadGEEToolbox==1.6.3
 
 Installing via Conda
 --------------------
 
-To install ``RadGEEToolbox`` version 1.6.1 using conda-forge (recommended: use a virtual environment)::
+To install ``RadGEEToolbox`` version 1.6.3 using conda-forge (recommended: use a virtual environment)::
 
     conda install conda-forge::radgeetoolbox
 
@@ -44,4 +44,4 @@ To verify that ``RadGEEToolbox`` was installed correctly::
 
     python -c "import RadGEEToolbox; print(RadGEEToolbox.__version__)"
 
-You should see ``1.6.1`` printed as the version number.
+You should see ``1.6.3`` printed as the version number.
