+++ README.md
@@ -2,7 +2,7 @@
 
 # RadGEEToolbox 🛠
 
-![](image-1.png)
+![](ReadMeCover.png)
 
 ### 🌎 Streamlined Multispectral & SAR Analysis for Google Earth Engine Python API
 
@@ -11,6 +11,69 @@
 **RadGEEToolbox** is an open-source Python package that simplifies the processing and analysis of satellite imagery using the Google Earth Engine Python API. It provides ready-to-use tools for filtering, masking, mosaicking, spectral index calculations, and extracting statistics from multispectral (Landsat, Sentinel-2) and SAR (Sentinel-1) datasets.
 
 Designed for both new and advanced users of Google Earth Engine, RadGEEToolbox minimizes repetitive scripting, accelerates common remote sensing workflows, and aims to maximize efficiency within the constraints of the Google Earth Engine API. Whether you’re building a time series of vegetation indices or extracting surface properties along transects, this package helps get results faster.
+
+_________
+## Getting Started with Google Earth Engine
+
+**RadGEEToolbox requires [access](https://developers.google.com/earth-engine/guides/access) to [Google Earth Engine (GEE)](https://earthengine.google.com/) and proper [authentication](https://developers.google.com/earth-engine/guides/auth) with the Python API.**
+
+For more details, see the official [Google Earth Engine Python API Getting Started Guide](https://developers.google.com/earth-engine/guides/quickstart_python) or [Python Installation Guide](https://developers.google.com/earth-engine/guides/python_install).
+
+
+### 1. Sign Up for Earth Engine
+Apply for access here: [https://earthengine.google.com/signup/](https://earthengine.google.com/signup/) using a Google account. Approval typically takes a few days.
+
+### 2. Install the Earth Engine Python API
+**The Earth Engine API is installed automatically when you install `RadGEEToolbox`.** However, if you would like to install manually:
+
+```bash
+pip install earthengine-api
+```
+
+### 3. Authenticate & Initialize
+Prior to using the Earth Engine Python client library, you need to authenticate and use the resultant credentials to initialize the Python client. The authentication flows use Cloud Projects to authenticate, and they're used for unpaid (free, noncommercial) use as well as paid use.
+
+
+**Run the following during your first use:**
+```python
+import ee
+ee.Authenticate()
+```
+***NOTE: Your GEE credentials will not permanantly be stored on your PC and you will periodically need to re-run `ee.Authenticate()` when `ee.Initialize()` returns an authentication error.***
+
+`ee.Autheticate()` will select the best authentication mode for your environment, and prompt you to confirm access for your scripts. To initialize, you will need to provide a project that you own, or have permissions to use. This project will be used for running all Earth Engine operations:
+
+```python
+ee.Initialize(project='my-project')
+```
+Replacing 'my-project' with the name of the Google Cloud Project you created on sign-up or any Google Cloud Project that has the GEE API enabled.
+
+### 4. Authentication Best Practices
+It is reccomended to use the following authentication procedure once you have completed your initial authentication:
+```python
+import ee
+
+try:
+    ee.Initialize()
+except Exception as e:
+    ee.Authenticate()
+    ee.Initialize()
+```
+
+### 5. Troubleshooting
+`AttributeError: module 'ee' has no attribute 'Initialize'`
+
[…]
