### 2025-06-04 1e8c2ac
hopeful fix for RTD examples page indexes
