### 2025-06-04 a80e2b6
added readme example to rtd example page
