### 2025-06-05 d80d6cf
fixing python dependency to >3.8 and small fixes to example notebooks
