+++ RadGEEToolbox/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.6.3"
+__version__ = "1.6.4"
 
 from .CollectionStitch import CollectionStitch, MosaicByDate
 from .GetPalette import get_palette
