+++ README.md
@@ -101,15 +101,15 @@ _____________
 
 ### Installing via pip
 
-To install `RadGEEToolbox` version 1.6.3 using pip (NOTE: it is recommended to create a new virtual environment):
+To install `RadGEEToolbox` version 1.6.4 using pip (NOTE: it is recommended to create a new virtual environment):
 
 ```bash
-pip install RadGEEToolbox==1.6.3
+pip install RadGEEToolbox==1.6.4
 ```
 
 ### Installing via Conda
 
-To install `RadGEEToolbox` version 1.6.3 using conda-forge (NOTE: it is recommended to create a new virtual environment):
+To install `RadGEEToolbox` version 1.6.4 using conda-forge (NOTE: it is recommended to create a new virtual environment):
 
 ```bash
 conda install conda-forge::radgeetoolbox
@@ -140,7 +140,7 @@ To verify that `RadGEEToolbox` was installed correctly:
 python -c "import RadGEEToolbox; print(RadGEEToolbox.__version__)"
 ```
 
-You should see `1.6.3` printed as the version number.
+You should see `1.6.4` printed as the version number.
 
 ### Want to Visualize Data? Install These Too
 
