+++ pyproject.toml
@@ -4,14 +4,14 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "RadGEEToolbox"
-version = "1.6.3"
+version = "1.6.4"
 description = "Streamlined Multispectral & SAR Analysis for Google Earth Engine Python API"
 authors = [
     {name = "Mark Radwin", email = "markradwin@gmail.com"}
 ]
 license = {text = "MIT"}
 readme = "README.md"
-requires-python = ">=3.6"
+requires-python = ">=3.8"
 dependencies = [
     "earthengine-api",
     "numpy",
