### 2025-06-05 e5bb7e7
update to min python version in readme
