### 2025-06-05 radwinskis/RadGEEToolbox#2 @radwinskis
Hi @mcwhity,

Thank you again for your detailed feedback and for serving as a JOSS reviewer. I’ve carefully reviewed and addressed each of your comments. Here’s a summary of the updates:

1. The version number is now consistent across the README, PyPI, conda-forge, and ReadTheDocs.

2. The `__init__.py` file now defines `__version__`, so:

   ```bash
   python -c "import RadGEEToolbox; print(RadGEEToolbox.__version__)"
   ```

   correctly outputs the package version.

3. `conda install conda-forge::radgeetoolbox` now installs the latest version with all core dependencies.

   a. I’ve added notes in the README and example notebooks that visualizing data requires optional dependencies: `ipykernel`, `geemap`, and `matplotlib`. These are not included by default, as they are not central to the package’s functionality, but will likely be needed if using notebooks or interactive maps.

   b. I updated the `.yaml` environment file so users can install the package for any Python version ≥ 3.8.

   c. I also discovered the package fails to install for Python < 3.8, so I have raised the minimum required Python version accordingly.

4. I’ve tested installation using both `conda` and `pip` in fresh environments. Both methods result in successful installs with no errors. The package version is correctly reported, and example notebooks run as expected.

5. The README and ReadTheDocs now include a new “Getting Started with Earth Engine” section to help new users set up a GEE account and authenticate. I also clarified that newer GEE requirements mean `ee.Initialize()` may need a `project="project-ID"` argument. This change wasn’t previously obvious to me because my default GCP project is GEE-enabled, so `ee.Initialize()` worked without specifying a project. Thanks for catching that.

6. As mentioned above, I now clearly document that `ipykernel` is required for Jupyter notebooks and `geemap` is needed for interactive mapping. I’ve kept these out of the core dependency list to avoid unnecessary bloat—especially since `geemap` pulls in many packages. I’ve aimed to make this trade-off transparent throughout the docs and examples.

   a. Thank you for suggesting Docker containerization. I agree it would improve reproducibility and ease of setup. For now, I’ve focused on resolving the more critical issues you identified (e.g., versioning, compatibility, and documentation). I do plan to provide a Dockerfile and container image in a future update once the package has stabilized. I appreciate your understanding.

Please let me know if there’s anything else I should revise. I’m grateful for your feedback and help in improving this package!
