### 2025-06-05 radwinskis/RadGEEToolbox#1 @radwinskis
Hi @JohnVolk  and @McWhity,

Thank you both for your additional feedback and for serving as a JOSS reviewer!

I have gone through the example notebooks and inserted screenshots or numerical outputs of sections where it would be helpful to see the output. For the notebooks which outline the basics of the package (without visualization or data aggregation) I make note of the example notebooks which do provide examples of data visualization.

Regarding the example images in the README "Usage Example", I didn't include the code for how I made those images as I was trying to keep the usage example short, highlighting the ways `RadGEEToolbox` is helpful for these workflows without importing other, optional packages. However, to fill this gap, I have included another notebook in the Example Notebooks directory, [Complete_ReadMe_Example.ipynb](https://github.com/radwinskis/RadGEEToolbox/blob/main/Example%20Notebooks/Complete_ReadMe_Example.ipynb), which is an extended version of what is shown in the README, providing the code for how the Lake Powell map and water area time series plot are created. I have provided a link to this notebook in the "Usage Example" section of the README for quick access. The typo with the extra bracket has also been fixed, and I now clarify the necessity for specifying `project=` within `ee.Initialize(project='project-ID')`  (see my comments in the [other issue](https://github.com/radwinskis/RadGEEToolbox/issues/2) opened by @McWhity for more information on improvements to documentation regarding Earth Engine authentication)

Thanks again for your thoughtful suggestions—they’ve helped make this package stronger and more user-friendly!
