### 2025-06-27 radwinskis/RadGEEToolbox#3 @sebhaan
Automated tests
While working on the JOSS review, I've noticed the repository does not appear to include any automated tests. I would recommend to add a test suite that covers the core functionality of RadGEEToolbox. Consider implementing:

- Unit tests for key methods in each module (LandsatCollection, Sentinel1Collection, Sentinel2Collection)
- Integration tests that verify Earth Engine API interactions work correctly
- Tests for spectral index calculations, masking operations, and statistical functions
