+++ paper/paper.bib
@@ -9,6 +9,23 @@
     Keywords = {Google Earth Engine, Remote sensing, Big data, Cloud computing, Geospatial analysis},
 }
 
+@article{Wu:2020, doi = {10.21105/joss.02305}, 
+url = {https://doi.org/10.21105/joss.02305}, year = {2020}, 
+publisher = {The Open Journal}, volume = {5}, number = {51}, 
+pages = {2305}, author = {Qiusheng Wu}, 
+title = {geemap: A Python package for interactive mapping with Google Earth Engine}, 
+journal = {Journal of Open Source Software} }
+
+@article{Montero:2021, 
+    doi = {10.21105/joss.03168}, 
+    url = {https://doi.org/10.21105/joss.03168}, 
+    year = {2021}, publisher = {The Open Journal}, 
+    volume = {6}, number = {62}, 
+    pages = {3168}, author = {David Montero}, 
+    title = {eemont: A Python package that extends Google Earth Engine}, 
+    journal = {Journal of Open Source Software} 
+}
+
 @article{Boucher:2018,
     url={https://esajournals.onlinelibrary.wiley.com/doi/10.1002/eap.1708},
     Author = {J. {Boucher} and K. {Weathers}, and H. {Norouzi}, and B. {Steele}},
@@ -79,3 +96,10 @@
     Year = {2020}
 }   
 
+@misc{geetools,
+  author = {gee-community},
+  title = {geetools},
+  year = {2017},
+  howpublished = {\url{https://github.com/gee-community/geetools}},
+  note = {Accessed: 2024-06-04}
+}
