### 2025-06-30 radwinskis/RadGEEToolbox#3 @sebhaan
Hi @radwinskis ,

Thank you for the detailed explanation and context about the practical limitations of automated testing when interactive authentication is required. I really appreciate how you've structured the extensive Jupyter notebooks as integration tests.

Your approach makes a lot of sense given the GEE service constraints, and the notebooks provide immediate visual feedback that traditional unit tests wouldn't offer. This is particularly valuable for a geospatial package where seeing the actual results (maps, statistics, etc.) is often more meaningful than abstract test assertions.

Interestingly, I noticed that [geemap's test suite](https://github.com/gee-community/geemap/tree/master/tests) uses mock testing with fake_ee.py modules that stub out Earth Engine API calls for their core unit tests. This allows them to test UI logic and utility functions without authentication, but your notebook approach actually provides better end-to-end validation of real Earth Engine interactions.

For future reference, you could potentially add some lightweight unit tests for utility functions that don't require GEE authentication (like parameter validation), but your current notebook-based integration testing approach is well-suited to the problem domain and provides excellent user-facing documentation simultaneously.

Thanks for taking the time to implement such thorough examples. Regarding the JOSS review, I'm satisfied with the current testing approach.
