### 2025-07-18 #8164 @JohnVolk
@elbeejay and @radwinskis Here are some comments and suggestions related to the scholarly effort redundancy, and statement of need, and automated tests.

**1. Automated testing should be added**

I found no evidence of automated testing. To strengthen the submission, the authors should include a basic test suite (e.g., for collection filtering, index calculations), configure pytest (possibly via pytest.ini), and add a workflow (e.g., GitHub Actions) to run tests on every commit.

**2. Academic value / redundancy suggestions**

The paper mentions that RadGEEToolbox is not a general-purpose Earth Engine extension or visualization tool but instead provides streamlined pipelines for common scientific workflows, including SAR preprocessing. This is a valid and mostly accurate claim. However, with regards to the potential overlap and differentiation with the other EE Python tools that have similar functionality, specifically geemap, eemont, and geetools- I suggest doing one or more of the following:

- Insert a brief comparison table or matrix summarizing the core features of RadGEEToolbox vs. geemap, eemont, and geetools. This would help clarify functional overlap vs. novelty. It could be inserted into the paper itself and also in the DOCs perhaps in a FAQ or as it's own section.

- Provide an example use-case where RadGEEToolbox simplifies or enables a workflow that would be more verbose or unavailable in other packages (e.g., SAR multilooking, swath mosaicking, zonal stats over NDWI time series).

- Make the SAR support more prominent in the novelty argument, as this is a key contribution that is currently not offered in the other GEE Python tools.

Otherwise, my review is going well and am not having any issues with the setup and code.
