### 2025-07-20 #8164 @JohnVolk
Just a few more thoughts to wrap up my review. Overall I think the software is well designed and is useful for EE beginners and specifically those interested in the common workflows with the Landsat and Sentinel data. The object oriented methods and structure are well thought out and clean.

Although it is very clear that the software is only wrapping the specified EE collections, it would be useful to know if there is any plans to extend the code to make it other GEE image collections (e.g. MODIS, VIIRS, ECOSTRESS, etc.). Could a generic wrapper class be designed or would this be a lot of work? This does not have to be added to the package, I am just wondering and if you have any plans for this it might be useful to add it on the contributing page as a potential for the community to assist.

Minor paper edit: Line 84: type "whererever possible"

As mentioned above in my last post, I still think that automated tests and distinguishing how it differs and is useful compared to the other packages would be really useful. Once those are addressed, what will conclude my review!
