### 2025-07-23 #8164 @radwinskis
Hi @elbeejay thank you for the ping. I have read the feedback from @JohnVolk and @McWhity and am appreciative of their thoughtful and detailed feedback. Things are a bit hectic for me at the moment but I hope to find time to address the noted points within the next 2-3 weeks. I hope that timeline works for everyone. Once again, I greatly appreciate the detailed feedback - I feel this review process has immensely improved this package.
