### 2025-08-11 365279d
review edits and backup pre-black linting
