+++ paper/paper.bib
@@ -59,7 +59,7 @@ journal = {Journal of Open Source Software} }
 @article{Radwin&Bowen:2024,
     url = {https://giw.utahgeology.org/giw/index.php/geosites/article/view/134},
     Author = {M. {Radwin} and B. {Bowen}},
-    Doi = {https://doi.org/10.31711/ugap.v51i.134},
+    Doi = {10.31711/ugap.v51i.134},
     Journal = {Geosites},
     Title = {Evolution of Great Salt Lake’s Exposed Lakebed (1984-2023): Variations in Sediment Composition, Water, and Vegetation from Landsat OLI and Sentinel MSI Satellite Reflectance Data},
     Volume = {51},
