+++ docs/example_notebooks.rst
@@ -1,8 +1,8 @@
-Examples
-========
+Example Notebooks
+=================
 
 To assist in getting you off the ground with using RadGEEToolbox, we have provided a number of example notebooks and will be adding more over time. 
-These notebooks are designed to demonstrate the functionality of RadGEEToolbox and provide a starting point for your own analyses.
+These notebooks are designed to demonstrate the functionality of RadGEEToolbox and provide a starting point for your own analyses. The order of notebooks below is intended to provide a logical progression from basic to more advanced usage.
 
 .. toctree::
    :maxdepth: 1
