### 2025-08-12 65eccf4
1.6.5 new features, JOSS paper & docs updates, and linted modules
