+++ paper/paper.md
@@ -20,7 +20,7 @@ bibliography: paper.bib
 
 # Summary
 
-`RadGEEToolbox` is an open-source Python package for processing and analyzing multispectral and Synthetic Aperture Radar (SAR) imagery using `Google Earth Engine` [@Gorelick:2017]. It supports Landsat TM/OLI, Sentinel-1 SAR, and Sentinel-2 MSI datasets, providing a modular, streamlined workflow for landscape change detection, environmental monitoring, and land classification. It reduces the need for repetitive code while promoting reproducibility and efficiency.
+`RadGEEToolbox` is an open-source Python package for processing and analyzing multispectral and Synthetic Aperture Radar (SAR) imagery using `Google Earth Engine` [@Gorelick:2017]. It supports Landsat TM/OLI, Sentinel-1 SAR, and Sentinel-2 MSI datasets, with plans to support additional platforms in the future (MODIS, VIIRS, etc). `RadGEEToolbox` provides a modular, streamlined workflow for landscape change detection, environmental monitoring, and land classification. It reduces the need for repetitive code while promoting reproducibility and efficiency.
 
 Features include subsetting, mosaicking, cloud/water masking, spectral index calculations, speckle filtering, backscatter conversions, spatial statistics, and integration of custom Earth Engine functions.
 
@@ -28,7 +28,26 @@ Features include subsetting, mosaicking, cloud/water masking, spectral index cal
 
 Working in `Google Earth Engine` often involves steep learning curves and complex scripting. `RadGEEToolbox` simplifies this by automating common multispectral and SAR workflows while remaining fully compatible with native GEE workflows. It's designed for students, researchers, and analysts working with Landsat and Sentinel imagery, enabling efficient time series analysis, surface property extraction, and visualization of environmental indicators. Ultimately, RadGEEToolbox makes it easier to perform multi-temporal analyses, extract surface properties or time series, generate land cover statistics, and visualize key environmental indicators by handling much of the complexity inherent in Google Earth Engine API interactions and image processing tasks.
 
-While several Python packages extend Google Earth Engine capabilities, including geemap for interactive mapping [@Wu:2020], eemont for simplified spectral index calculations [@Montero:2021], and geetools [@geetools] for broad utility functions, RadGEEToolbox takes a different approach. Rather than providing a general-purpose set of method extensions to Earth Engine objects, RadGEEToolbox offers a modular, workflow-oriented API centered around user-defined collection classes in addition to a set of Earth Engine method extensions. These objects bundle common operations, such as filtering, cloud masking, mosaicking, spectral index calculation, and zonal statistics into streamlined, reusable workflows with minimal boilerplate code. This design emphasizes rapid implementation of end-to-end pipelines for scientific studies, particularly for hydrology, geomorphology, forestry, and time series landscape analysis. In this way, RadGEEToolbox complements existing packages while providing a focused solution for efficient, reproducible remote sensing workflows.
[…]
