+++ docs/example_notebooks.rst
@@ -3,6 +3,7 @@ Example Notebooks
 
 To assist in getting you off the ground with using RadGEEToolbox, we have provided a number of example notebooks and will be adding more over time. 
 These notebooks are designed to demonstrate the functionality of RadGEEToolbox and provide a starting point for your own analyses. The order of notebooks below is intended to provide a logical progression from basic to more advanced usage.
+If you are interested in learning about how to visualize the images you create with RadGEEToolbox, please see the `Complete_ReadMe_Example.ipynb` or `Palettes_and_Visualization.ipynb` notebook.
 
 .. toctree::
    :maxdepth: 1

[…]
