### 2025-08-12 0e31720
v1.6.6 updates band scaling fix
