+++ README.md
@@ -120,15 +120,15 @@ _____________
 
 ### Installing via pip
 
-To install `RadGEEToolbox` version 1.6.5 using pip (NOTE: it is recommended to create a new virtual environment):
+To install `RadGEEToolbox` version 1.6.6 using pip (NOTE: it is recommended to create a new virtual environment):
 
 ```bash
-pip install RadGEEToolbox==1.6.5
+pip install RadGEEToolbox==1.6.6
 ```
 
 ### Installing via Conda
 
-To install `RadGEEToolbox` version 1.6.5 using conda-forge (NOTE: it is recommended to create a new virtual environment):
+To install `RadGEEToolbox` version 1.6.6 using conda-forge (NOTE: it is recommended to create a new virtual environment):
 
 ```bash
 conda install conda-forge::radgeetoolbox
@@ -159,7 +159,7 @@ To verify that `RadGEEToolbox` was installed correctly:
 python -c "import RadGEEToolbox; print(RadGEEToolbox.__version__)"
 ```
 
-You should see `1.6.5` printed as the version number.
+You should see `1.6.6` printed as the version number.
 
 ### Want to Visualize Data? Install These Too
 
