### 2025-08-12 #8164 @radwinskis
Hi @elbeejay , I have addressed the feedback from @McWhity and @JohnVolk – see responses below.
@McWhity and @JohnVolk, thank you again for your valuable reviews!
Regarding the comments from @McWhity  , I have followed the suggestions and:
-	Fixed GRD typo
-	Fixed the GitHub Discussions link
-	Added the Statement of Need to the RtD documentation
-	Fixed the duplicated doi link
-	Made note of the order of example notebooks in the RtD documentation
-	Added more information and interactive maps to the example notebooks where I introduce the different tiling systems for Landsat and Sentinel
-	Added an additional example of retrieving image collection metadata
-	Checked that the masking cloud section is duplicated. I found no duplication, but rather one is for masking clouds and the other is for masking water – but their structure is very similar so at a first glance it looks like a duplicated cell.
-	Decided not to add optional visualizations to the `Image_Collections_&_Useful_Operations.ipynb` + `Multispectral_Functions_Examples.ipynb` notebooks as I do not want to further bloat the notebooks – however, I have added more notices throughout the notebook and documentation regarding which notebooks to view for examples of visualizing the products. My intention for these notebooks is to mainly provide a comprehensive list of examples of how to use the package by demonstrating the various features included in each module. I understand it may be a bit too technical, so I am planning on adding an additional example notebook (or multiple) in the coming weeks/months which will show a handful of remote sensing workflows and visualizations – being less technical and more practical for newer GEE or Python users.
-	Importantly, I implemented linting with black top clean up the code and shorten the long lines. I am a big fan of these improvements – thank you for the suggestion.

Regarding the comments from @JohnVolk  :
-	Automated testing. This issue was also brought up by @Sebhaan and although I agree it would be best to include automatic testing, this is very challenging with GEE as any API requests will need GEE authentication and thus requires either exposing keys or manually authenticating when updating the package for automatic tests. Thus, I have opted to use the Example Notebooks as practical tests instead of implementing automatic tests. After updating the package, I run any of the example notebooks to ensure the update has not broken functionality. See my reply to @Sebhaan: https://github.com/radwinskis/RadGEEToolbox/issues/3
-	Your point regarding the academic value of RadGEEToolbox is very valid and I have gone ahead and included a comparison table between RadGEEToolbox, eemont, and geetools to the ReadMe, RtD docs, and JOSS paper – which highlights the unique features of RadGEEToolbox (such as SAR support) and overlap with the other packages, and supports the need of RadGEEToolbox. I have also added a few sentences in the paper and documentation regarding the purpose of RadGEEToolbox. I hope this is sufficient. In the future I plan on creating some side-by-side code examples for specific tasks using RadGEEToolbox, vanilla GEE API, and eemont to further illustrate how RadGEEToolbox greatly simplifies remote sensing workflows for Landsat and Sentinel platforms.
-	I have added to the paper that I do have plans to expand the functionality of RadGEEToolbox to include modules for other platforms like MODIS and VIIRS. To respond to your other point, I do believe it would be difficult to create a generic wrapper that works for all platforms hosted on GEE – but perhaps it is something I might explore in the future! My main focus has been on supporting high resolution platforms that are most common in applied remote sensing analyses for geological and environmental studies (which supports my own research), hence the support for Landsat, Sentinel-1, and Sentinel-2 – but I agree that this package would be more valuable with expanded support. I have noted in the paper and documentation that I plan on continuously expanding the functionality of the package, and invite any suggestions from end users.
-	I have fixed the typo mentioned on line 84.
