### 2025-08-12 #8164 @JohnVolk
Hi @radwinskis thanks for addressing these issues, I mentioned how you may go about automated tests with GEE authentication with GitHub actions on your [issues](https://github.com/radwinskis/RadGEEToolbox/issues/3). I will defer to @elbeejay to decide if that is needed.

I think the Table 1 you added to the paper does a lot to address my concern of potential overlap and helps highlight the contribution of this package as compared to some of the more related GEE Python packages that are out there. A couple suggestions though for the table: 1) it is hard to read the table due to the word wrapping of the first column; 2) you may want to check and switch geetools from a NO to a YES for zonal stats and time series zonal stats, see https://geetools.readthedocs.io/en/stable/usage/reduce.html#reduce-over-muliple-regions as an example that does compute time series zonal stats if I understand it correctly.

Other than that, I am satisfied, good job, and it's great to hear you have plans to include new image collections like VIIRS and MODIS!
