### 2025-08-15 #8164 @elbeejay
I appreciate the effort all around to have a productive discussion around the review comments. I've chimed in on https://github.com/radwinskis/RadGEEToolbox/issues/3. @radwinskis, if you could double-check and close the other open issues pertaining to this JOSS review that you've addressed (which I believe is all of them) that would be great.

I agree with your comment regarding the table formatting @JohnVolk - @openjournals/dev is there a way for the author to control the spacing of columns in a table?

@McWhity - hoping you'll be able to revisit this review soon as @radwinskis has made a number of revisions and hopefully addressed your review comments.
