### 2025-08-15 #8164 @arfon
> I agree with your comment regarding the table formatting @JohnVolk - @openjournals/dev is there a way for the author to control the spacing of columns in a table?

Not that I'm aware of sorry. You might be able to use a LaTeX table to achieve this though.
