### 2025-08-15 radwinskis/RadGEEToolbox#3 @radwinskis
Hi @JohnVolk and @elbeejay ,

Thanks for the information @JohnVolk ! Yesterday I was looking more into this and have started working on setting up automatic tests. It is much more straight forward than I anticipated, and I'm excited about being able to include automatic testing. I didn't completely finish the setup for the testing workflow, but I will likely have it completed early next week in addition to handling the other comments. Thanks for the suggestions and pointing me in the right direction!
