### 2025-08-15 #8164 @JohnVolk
@elbeejay and @radwinskis Replying about the table spacing/word wrapping in Table 1 in the paper.md. It's not a big blocker and I don't want to hold up the paper, it is just for aesthetics.

You might want to try using backslashes like ``Zonal\ Time-series\ Extraction`` to prevent word wrapping in the first column. If that doesn't work, I don't know if you can use HTML in the tables to prevent wrapping or something else (LaTeX?). Sorry I can't be of more help. But if you want to test a few things it looks like you can compile the PDF  locally: https://joss.readthedocs.io/en/latest/paper.html#locally
