### 2025-08-16 radwinskis/RadGEEToolbox#3 @elbeejay
Appreciate the additional information @JohnVolk - my question about cost was more about the GEE side of having and using those credentials than the GitHub storing and usage of secrets. Although it sounds like it is not going to be a problem for @radwinskis.
