### 2025-08-19 a1c5ffe
Create autotest.yml
Adding workflow for automatic testing
