+++ .github/workflows/autotest.yml
@@ -1,0 +1,37 @@
+name: CI
+
+on:
+  push:
+  pull_request:
+  workflow_dispatch:
+
+permissions:
+  contents: read
+
+jobs:
+  tests:
+    runs-on: ubuntu-latest
+
+    # Expose your repo secrets to the job (they'll be empty on forked PRs)
+    env:
+      GEE_SA_KEY: ${{ secrets.GEE_SA_KEY }}
+      GEE_PROJECT: ${{ secrets.GEE_PROJECT }}
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
+          cache: "pip"
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -e .[dev]
+          pip install earthengine-api pytest
+
+      - name: Run tests
+        run: pytest -q
