### 2025-08-19 2fce081
added tests and adjusting feature comparison table
