+++ paper/paper.md
@@ -36,7 +36,7 @@ This design emphasizes rapid implementation of end-to-end pipelines for scientif
 **Table 1.** Comparison of functionality between RadGEEToolbox, eemont, and geetools.
 
 | Capability | **RadGEEToolbox** | **eemont** | **geetools** |
-|---|:---:|:---:|:---:|
+|:--------------:|:---:|:---:|:---:|
 | **Dataset & Workflow Specific API's** | **YES** | NO | NO |
 | **Synthetic Aperture Radar (S1) Support** | **YES** | NO | NO |
 | **Zonal Time-series Extraction** | **YES** | **YES** | **YES** |
