### 2025-08-19 radwinskis/RadGEEToolbox#3 @radwinskis
As of the most recent commit, RadGEEToolbox now implements automatic testing thanks to the help of @JohnVolk. Core functionality is currently tested and completes rapidly, and more tests will be added in the near future to perform comprehensive testing of nearly all features. I will now be closing this issue.
