### 2025-08-19 #8164 @JohnVolk
Hi @radwinskis that sounds good, it looks much better and I'm glad to hear the testing was implemented and will be expanded! Just one thing- do you think that geetools also does the area time series extraction. It seems like that is what is described here: https://geetools.readthedocs.io/en/stable/usage/reduce.html#reduce-over-muliple-regions Sorry if it was not clear before.
