### 2025-08-19 #8164 @McWhity
@radwinskis
One minor point: the line-length reduction in the Python code looks good. However, you could do the same for the documentation docstrings. It’s a style consideration and can be a bit more challenging, but it’s worth keeping in mind where applicable.
