### 2025-08-24 #8164 @elbeejay
@radwinskis, apologies for the delay on my end. The paper looks to be well-written. Formatting-wise, I'd ask that you please go through and double-check some of the typesetting, particularily around the use of the backticks. My recommendation is to use the backticks around Python packages, modules and objects only. I would remove the backticks around "Google Earth Engine" and ensure they exist around references to the `RadGEEToolbox`.

The bot will generate "post-review checklist" below, and I will need your input and confirmation on the software version to use for publication (looks like 1.6.6), the link to the archived copy, etc.
