### 2025-08-25 #8164 @radwinskis
Hi @elbeejay, I have updated the paper to fix the use of backticks per your feedback.

For the software release to be used in the paper, version 1.6.7 would be best as it includes the automatic test files and other recent updates.

The release (v1.6.7) is archived using Zenodo through the following DOI link: https://doi.org/10.5281/zenodo.16943095

I have verified that the archive states the correct title, author list (including ORCID), version tag, and license.
