### 2025-09-03 #8164 @elbeejay
@radwinskis looking for a minor edit, see below:

> [@radwinskis](https://github.com/radwinskis) I'd also suggest removing the line:
>
> ```
>     equal-contrib: true
> ```
>
> from the paper MarkDown file as you are the sole author, so I don't think you need the note that you contributed equally to be included in the PDF.
