### 2025-09-03 978ea7f
removed equal contribution line
