### 2025-09-03 #8164 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
