### 2025-09-17 8295a9d
References capitalization fix for .bib
