+++ paper/paper.bib
@@ -3,7 +3,7 @@
     Author = {N. {Gorelick} and M. {Hancher} and M. {Dixon} and S. {Ilyushchenko} and D. {Thau} and L. {Moore}},
     Doi = {10.1016/j.rse.2017.06.031},
     Journal = {Remote Sensing of Environment},
-    Title = {Google Earth Engine: Planetary-scale geospatial analysis for everyone},
+    Title = {{Google Earth Engine}: Planetary-scale geospatial analysis for everyone},
     Volume = {202},
     Year = {2017},
     Keywords = {Google Earth Engine, Remote sensing, Big data, Cloud computing, Geospatial analysis},
@@ -13,7 +13,7 @@
 url = {https://doi.org/10.21105/joss.02305}, year = {2020}, 
 publisher = {The Open Journal}, volume = {5}, number = {51}, 
 pages = {2305}, author = {Qiusheng Wu}, 
-title = {geemap: A Python package for interactive mapping with Google Earth Engine}, 
+title = {geemap: A Python package for interactive mapping with {Google Earth Engine}}, 
 journal = {Journal of Open Source Software} }
 
 @article{Montero:2021, 
@@ -22,7 +22,7 @@ journal = {Journal of Open Source Software} }
     year = {2021}, publisher = {The Open Journal}, 
     volume = {6}, number = {62}, 
     pages = {3168}, author = {David Montero}, 
-    title = {eemont: A Python package that extends Google Earth Engine}, 
+    title = {eemont: A Python package that extends {Google Earth Engine}}, 
     journal = {Journal of Open Source Software} 
 }
 
@@ -31,7 +31,7 @@ journal = {Journal of Open Source Software} }
     Author = {J. {Boucher} and K. {Weathers}, and H. {Norouzi}, and B. {Steele}},
     Doi = {10.1002/eap.1708},
     Journal = {Ecological Applications},
-    Title = {Assessing the effectiveness of Landsat 8 chlorophyll a retrieval algorithms for regional freshwater monitoring},
+    Title = {Assessing the effectiveness of {Landsat 8} chlorophyll a retrieval algorithms for regional freshwater monitoring},
     Volume = {28},
     Year = {2018}
 }
@@ -41,7 +41,7 @@ journal = {Journal of Open Source Software} }
     Author = {W. {Buma} and S. {Lee}},
     Doi = {10.3390/rs12152437},
     Journal = {Remote Sensing},
-    Title = {Evaluation of Sentinel-2 and Landsat 8 Images for Estimating Chlorophyll-a Concentrations in Lake Chad, Africa},
+    Title = {Evaluation of {Sentinel-2} and {Landsat 8} Images for Estimating Chlorophyll-a Concentrations in {Lake Chad}, {Africa}},
     Volume = {12},
     Year = {2020}
 }
@@ -51,7 +51,7 @@ journal = {Journal of Open Source Software} }
     Author = {M. {Radwin} and B. {Bowen}},
     Doi = {10.1002/esp.5089},
     Journal = {Earth Surface Processes and Landforms},
-    Title = {Mapping mineralogy in evaporite basins through time using multispectral Landsat data: Examples from the Bonneville basin, Utah, USA},
+    Title = {Mapping mineralogy in evaporite basins through time using multispectral {Landsat} data: Examples from the {Bonneville} basin, {Utah}, {USA}},
     Volume = {46},
     Year = {2021}
 }
@@ -61,7 +61,7 @@ journal = {Journal of Open Source Software} }
     Author = {M. {Radwin} and B. {Bowen}},
     Doi = {10.31711/ugap.v51i.134},
     Journal = {Geosites},
[…]
