### 2025-09-19 #8164 @kthyng
Congratulations on your new publication @radwinskis! Many thanks to editor @elbeejay and to reviewers @JohnVolk, @McWhity, and @sebhaan for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
