### 2025-09-19 #8164 @radwinskis
@kthyng thank you! Very happy about having completed this process. I did want to note that it seems there is an issue with the title of the paper: "RadGEEToolbox: Streamlined Multispectral &amp; SAR Analysis for Google Earth Engine" - specifically the "&amp;" portion which should just be "&". This appears in the title and citation. I hope this is an easy fix.

Thank you for your help!
